refactor helpers and separate into files, complete seed data for all undecided/closed requests
